straightened out some of the Dashboard conditional rendering, modified AdminSidebar (now DashboardSidebar) to render menu items conditionally, moved employees state up higher to avoid re-fetching everytime admin switches back and forth between views
