+++ frontend/src/Dashboard/AdminEmployeesTable/EmployeesListNew.jsx
@@ -4,28 +4,7 @@ import { listUsers } from '../../utils/api'
 import { Table } from 'flowbite-react'
 import EmployeeCardNew from './EmployeeCardNew'
 
-export default function EmployeesList() {
-    const [employees, setEmployees] = useState(null)
-    const [error, setError] = useState(null)
-
-    // Fetches all users from the API
-    const loadUsers = useCallback(async () => {
-        const abortController = new AbortController()
-
-        try {
-            const response = await listUsers(abortController.signal)
-            setEmployees(response)
-        } catch (error) {
-            setError(error)
-        } finally {
-            abortController.abort()
-        }
-    }, [])
-
-    useEffect(() => {
-        loadUsers()
-    }, [loadUsers])
-
+export default function EmployeesList({ employees, setEmployees }) {
     if (employees) {
         return (
             <>
@@ -36,8 +15,7 @@ export default function EmployeesList() {
                     >
                         <EmployeeCardNew
                             employee={employee}
-                            setError={setError}
-                            loadUsers={loadUsers}
+                            setEmployees={setEmployees}
                             imgNumber={employee.person_id}
                         />
                     </Table.Row>
