+++ frontend/src/Dashboard/Dashboard.jsx
@@ -4,14 +4,15 @@ import {
     readUserById,
     readAveragesById,
     readCompanyMetrics,
+    listUsers,
 } from '../utils/api'
 import Spinner from '../utils/Spinner'
 import ActivityLogModal from './ActivityLog/ActivityLogModal'
 import AdminProgressCharts from './AdminProgressCharts/AdminProgressCharts'
-import AdminSidebar from './AdminSidebar/AdminSidebar'
 import AdminEmployeesTable from './AdminEmployeesTable/AdminEmployeesTable'
 import UserRecordsTable from './UserRecordsTable/UserRecordsTable'
 import UserProgressCharts from './UserProgressCharts/UserProgressCharts'
+import DashboardSidebar from './DashboardSidebar/DashboardSidebar'
 
 export default function AdminHome() {
     const { userId } = useParams()
@@ -28,31 +29,44 @@ export default function AdminHome() {
         sleep_duration_total: 0,
         quality_of_sleep_average: 0,
     })
-    const [view, setView] = useState("user")
-
+    const [view, setView] = useState('user')
+    const [employees, setEmployees] = useState(null)
 
     // Fetches user from the API along with lastMonthAverages for that user and company-wide lastMonthAverages for admin purposes
     const loadData = useCallback(async () => {
         const abortController = new AbortController()
 
         try {
+            // Reads user designated by url param
             const readUserResponse = await readUserById(
                 userId,
                 abortController.signal
             )
+            setUser(readUserResponse)
+
+            // Reads that user's lastMonthAverages from his/her entries
             const readAveragesResponse = await readAveragesById(
                 userId,
                 abortController.signal
             )
             setAverages(readAveragesResponse)
-            setUser(readUserResponse)
+
             if (readUserResponse.admin) {
+                // Reads all user's lastMonthAverages and totals
                 const readCompanyMetricsResponse = await readCompanyMetrics(
                     abortController.signal
                 )
                 setCompanyMetrics(readCompanyMetricsResponse)
-                setView("admin")
+
+                // Lists all users for use in AdminEmployeesTable
+                const listUsersResponse = await listUsers(
+                    abortController.signal
+                )
+                setEmployees(listUsersResponse)
+
+                setView('admin')
             }
+
         } catch (error) {
             console.error(error)
         } finally {
@@ -76,40 +90,55 @@ export default function AdminHome() {
         setIsModalOpen(true)
     }
 
+
     if (renderConditionsMet()) {
-        if (view === "admin") {
+        if (view === 'admin') {
             return (
                 <>
                     <section className="bg-slate-100 py-5">
-                        <AdminProgressCharts companyMetrics={companyMetrics}/>
-                        <div className=" flex flex-row w-full mx-auto mt-5 max-w-5xl max-h-[50vh] rounded-lg shadow-md overflow-hidden ">
-                            <AdminSidebar openModal={openModal} />
-                            <AdminEmployeesTable />
+                        {view === 'admin' && (<AdminProgressCharts companyMetrics={companyMetrics} />)}
+                        <div className="flex flex-row w-full mx-auto mt-5 max-w-5xl max-h-[50vh] rounded-lg shadow-md overflow-hidden ">
+                            <DashboardSidebar
+                                openModal={openModal}
+                                userIsAdmin={user.admin}
+                                view={view}
+                                setView={setView}
+                            />
+                            {view === 'admin' ? (<AdminEmployeesTable employees={employees} setEmployees={setEmployees} />) : (<UserRecordsTable userId={userId} />)}
                         </div>
                     </section>
-                    {isModalOpen && <ActivityLogModal setIsModalOpen={setIsModalOpen}/>}
+                    {isModalOpen && (
+                        <ActivityLogModal setIsModalOpen={setIsModalOpen} />
+                    )}
                 </>
             )
         } else {
             return (
                 <>
                     <section className="bg-slate-100 py-5">
-                        <UserProgressCharts averages={averages}/>
+                        <UserProgressCharts averages={averages} />
                         <div className="flex flex-row w-full mx-auto mt-5 max-w-5xl max-h-[50vh] rounded-lg shadow-md overflow-hidden ">
-                            <AdminSidebar openModal={openModal} />
+                            <DashboardSidebar
+                                openModal={openModal}
+                                userIsAdmin={user.admin}
+                                view={view}
+                                setView={setView}
+                            />
                             <UserRecordsTable userId={userId} />
                         </div>
                     </section>
-                    {isModalOpen && <ActivityLogModal setIsModalOpen={setIsModalOpen}/>}
+                    {isModalOpen && (
+                        <ActivityLogModal setIsModalOpen={setIsModalOpen} />
+                    )}
                 </>
             )
         }
     } else {
         return (
             <>
-            <div className='py-20'>
-                <Spinner />
-            </div>
+                <div className="py-20">
+                    <Spinner />
+                </div>
             </>
         )
     }
